upload an image file for a food type

+++ src/Controller/AdminTypeController.php
@@ -7,9 +7,11 @@ use App\Form\TypeType;
 use App\Repository\TypeRepository;
 use Doctrine\Persistence\ManagerRegistry;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\HttpFoundation\File\Exception\FileException;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\String\Slugger\SluggerInterface;
 
 class AdminTypeController extends AbstractController
 {
@@ -32,6 +34,7 @@ class AdminTypeController extends AbstractController
     /**
      * Create or Update a type
      *
+     * @param $slugger
      * @param Type|null $type
      * @param Request $request
      * @param ManagerRegistry $managerRegistry
@@ -39,7 +42,7 @@ class AdminTypeController extends AbstractController
      */
     #[Route('/admin/type/create', name: 'app_create_type', methods: 'GET|POST')]
     #[Route('/admin/type/{id}', name: 'app_update_type', methods: 'GET|POST')]
-    public function updateOrCreateType(Type $type = null, Request $request, ManagerRegistry $managerRegistry): Response
+    public function updateOrCreateType(SluggerInterface $slugger, Type $type = null, Request $request, ManagerRegistry $managerRegistry): Response
     {
         if(!$type) {
             $type = new Type();
@@ -48,6 +51,31 @@ class AdminTypeController extends AbstractController
         $form = $this->createForm(TypeType::class, $type);
         $form->handleRequest($request);
         if($form->isSubmitted() && $form->isValid()) {
+
+            // Get the uploaded image file
+            $imageFile = $form->get('imageFile')->getData();
+            // Check if the image file is valid
+            if($imageFile){
+                $imageFileOriginal = pathinfo($imageFile->getClientOriginalName(), PATHINFO_FILENAME);
+                // Reformat the image file name to be conform to an URL (without special chars) with Slugger interface
+                $imageFileReformat = $slugger->slug($imageFileOriginal);
+                // Create an unique name & unique id for the image file
+                $imageName = $imageFileReformat.'-'.uniqid().'-'.$imageFile->getExtension();
+
+                // Move the image file to a specific directory in the server
+                try {
+                    $imageFile->move(
+                        $this->getParameter('repertoire_images_type'),
+                        $imageName
+                    );
+                } catch(FileException $e) {
+                    throw $e;
+                }
+
+                // Save the name of the image file only
+                $type->setImage($imageName);
+            }
+
             $isUpdating = $type->getId() !== null;
             $managerRegistry->getManager()->persist($type);
             $managerRegistry->getManager()->flush();
@@ -57,7 +85,8 @@ class AdminTypeController extends AbstractController
 
         return $this->render('admin/admin_type/adminCreateUpdate.html.twig', [
             "type" =>$type,
-            "form" => $form->createView()
+            "form" => $form->createView(),
+            "isUpdating" => $type->getId() !== null
         ]);
     }
 
@@ -72,13 +101,15 @@ class AdminTypeController extends AbstractController
     #[Route('/admin/type{id}', name: 'app_delete_type', methods: 'DEL')]
     public function deleteType(ManagerRegistry $managerRegistry, Type $type, Request $request): Response
     {
-        $image = $this->getParameter('repertoire_images_type') . '/' . $type->getImage();
+        $image = $this->getParameter('repertoire_images_type').'/'. $type->getImage();
 
         if ($this->isCsrfTokenValid('REM' . $type->getId(), $request->get('_token'))) {
             $managerRegistry->getManager()->remove($type);
             $managerRegistry->getManager()->flush();
-            if (file_exists($image)) {
-                unlink($image);
+            if($image){
+                if (file_exists($image)) {
+                    unlink($image);
+                }
             }
             $this->addFlash('success', 'La suppression du type ' . $type->getName() . ' a bien été effectuée.');
             return $this->redirectToRoute('app_admin_type');
